exibindo mensagem de erro ao excluir um item

+++ app/Services/Type/TypeService.php
@@ -2,16 +2,25 @@
 
 namespace App\Services\Type;
 
+use App\Repositories\Transaction\TransactionRepositoryInterface;
 use App\Repositories\Type\TypeRepositoryInterface;
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Redirector;
 
 class TypeService implements TypeServiceInterface
 {
-    private $typeRepository;
+    protected TypeRepositoryInterface $typeRepository;
+    protected TransactionRepositoryInterface $transactionRepository;
 
-    public function __construct(TypeRepositoryInterface $typeRepository)
+    public function __construct(
+        TypeRepositoryInterface $typeRepository,
+        TransactionRepositoryInterface $transactionRepository
+    )
     {
         $this->typeRepository = $typeRepository;
+        $this->transactionRepository = $transactionRepository;
     }
 
     public function allTypes()
@@ -34,8 +43,16 @@ class TypeService implements TypeServiceInterface
         return $this->typeRepository->updateType($request, $id);
     }
 
-    public function deleteType(int $id)
+    public function deleteType(int $id): Redirector|RedirectResponse|Application
     {
-        return $this->typeRepository->deleteType($id);
+        $type = $this->transactionRepository->getTransactionByTypeId($id);
+
+        if ($type > 0) {
+            return back()->with('error', "Item não pode ser removido");
+        }
+
+        $this->typeRepository->deleteType($id);
+
+        return redirect("/types")->with('success', "Cadastro removido com Sucesso!");
     }
 }
